Puzzle2 is making progress - Always finds a solution, each run is unique
The agent "builds up" a tower starting with an empty one.

there are 3 mutations - add, remove, and swap: each has a probability (2/12) of happening. Add adds a block from "out" to the tower, remove takes a block from the tower and stores it in "out"
swap picks two indices on the most fit tower and swaps their blocks.